Move sleeping logic into separate step and plugin
There wasn't a good logical place for it, so I think a separate plugin
makes sense for now. It helps keep things modular as well.

It can be moved to a better place later if there is one.
